Refactoring metric package structure.
