+++ pom.xml
@@ -6,6 +6,7 @@
 
     <groupId>com.rdebokx</groupId>
     <artifactId>formica</artifactId>
+    <packaging>pom</packaging>
     <version>0.1.0-SNAPSHOT</version>
 
     <name>Formica</name>
@@ -119,6 +120,7 @@
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
             <version>4.12</version>
+            <scope>test</scope>
         </dependency>
     </dependencies>
